+++ src/info.js
